Update on nn_buildng_script/R
There are three main functions that make the building scores for neuralnet. Scores, new, and all, these should be fully functional. One thing that needs to be done is to incorporate CS_ID into the model's for new and all, otherwise this should be ready for integration less a combination function of all three main functions if needed.